+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@
       is raised
   * Fixed a small bug with some minimal preprocessing when getting value by the "key". 
     Now all the provided keys are passing the `preprocessor` to conform with the rules.
+* Ticket https://github.com/PandaHugMonster/py-simputils-config/issues/20
+  * Implemented `simputils.config.components.prisms.ObjConfigStorePrism` for `ConfigStore`.
+    It allows to use "Config Object Style Access" for accessing key/value pairs through `obj` field of `ConfigStore` object.
+  * And documentation for it: [Config Object Style Access](config-object-style-access.md)
 
 
 ## 1.0.5